Put non blockdev functions in blockdev.rs elsewhere
These functions all primarily interact with devices not as Stratis
blockdevs, but just as devices.

Most go into a new file, device.rs, but initialize() and resolve_devices()
are placed in blockdevmgr.rs. This is an appropriate spot because it is
about managing multiple devices to make them blockdevs.

Fixup an error message. Previously it was too specific and didn't get at the
actual problem. Also, shorten dev_info() error handling. Our errors promote
quite easily, so that works out nicely.
