+++ tests/real_tests.rs
@@ -18,7 +18,7 @@ use serde_json::{Value, from_reader};
 use self::devicemapper::{Bytes, Sectors};
 
 use libstratis::consts::IEC;
-use libstratis::engine::strat_engine::blockdev::wipe_sectors;
+use libstratis::engine::strat_engine::device::wipe_sectors;
 
 use util::logger::init_logger;
 use util::simple_tests::test_basic_metadata;
